refactor: build navbar functionality to share same code as mobile navbar
