+++ src/components/Navbar/components/LogoHomeButton.tsx
@@ -1,0 +1,10 @@
+import Logo from 'components/Logo'
+import Link from 'next/link'
+
+export const LogoHomeButton = () => (
+  <Link href="/">
+    <a>
+      <Logo />
+    </a>
+  </Link>
+)
